Moved Searchbar in own component
